cloud-functions/airtable-control/requirements.txt
firebase_admin
flask
python-dateutil
google-cloud
google-cloud-translate
